+++ fivefury/ymap/extensions.py
@@ -1,5 +1,0 @@
-from ..map_extensions import *
-from ..map_extensions import __all__ as _SHARED_EXTENSION_EXPORTS
-from .extension_defs import YMAP_EXTENSION_STRUCT_INFOS
-
-__all__ = [*_SHARED_EXTENSION_EXPORTS, "YMAP_EXTENSION_STRUCT_INFOS"]  # noqa: PLE0604